Removed some unnecessary spaces

+++ ViAn/Video/video_player.cpp
@@ -63,7 +63,6 @@ void video_player::run()  {
     int delay = (1000/frame_rate);
     capture.set(CV_CAP_PROP_POS_FRAMES, current_frame);
     while (!video_stopped && capture.read(frame)) {
-
         const clock_t begin_time = std::clock();
         convert_frame();
 
@@ -88,7 +87,6 @@ void video_player::run()  {
         }
         m_mutex->unlock();
     }
-
     video_stopped = true;
     capture.set(CV_CAP_PROP_POS_FRAMES, 0);
     emit update_current_frame(0);
@@ -253,14 +251,11 @@ bool video_player::set_current_frame_num(int frame_nbr) {
     if (frame_nbr >= 0 && frame_nbr < get_num_frames()) {
         // capture.set() sets the number of the frame to be read.
         if (video_paused) {
-            capture.set(CV_CAP_PROP_POS_FRAMES, (double)frame_nbr);
+            capture.set(CV_CAP_PROP_POS_FRAMES, frame_nbr);
         } else {
             set_new_frame = true;
             new_frame_num = frame_nbr;
         }
-        // capture.read() will read the frame and advance one step.
-        //capture.read(frame);
-
         return true;
     }
     return false;
